(#4) feature: add jwt error type
