--- src/main/kotlin/com/socksfactory/naeyangkku/global/auth/jwt/exception/type/JwtErrorType.kt
@@ -1,11 +1,0 @@
-package com.socksfactory.naeyangkku.global.auth.jwt.exception.type
-
-enum class JwtErrorType {
-    OK,
-    ExpiredJwtException,
-    SignatureException,
-    MalformedJwtException,
-    UnsupportedJwtException,
-    IllegalArgumentException,
-    UNKNOWN_EXCEPTION
-}
